Add try/except to account for index out of range error
For 32 bit Windows download, the file that had been searched to retrieve the version no longer contained that data so the regex was not finding any information.  The only other place I was able to find it was in the installer binary, which, while not ideal at least was preceded by "manifest version" which at least reads as important enough to be consistent.  It will only search that if the initial search fails, as I am unsure if this was an error on google's part that will be resolved in a future release or not.